Updated simulation parameters and added tests
